packages/python-services/requirements.txt
aiohttp==2.3.10
asn1crypto==0.24.0
async-lru==0.1.0
async-timeout==2.0.1
attrdict==2.0.0
backcall==0.1.0
base58==1.0.0
bumpversion==0.5.3
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coincurve==7.1.0
cryptography==2.2.2
cytoolz==0.9.0.1
decorator==4.3.0
Django==2.0.6
eth-abi==1.1.1
eth-account==0.2.3
eth-bloom==1.0.0
eth-hash==0.1.4
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-tester==0.1.0b26
eth-utils==1.0.3
ethereum==2.3.1
future==0.16.0
hexbytes==0.1.0
idna==2.7
idna-ssl==1.0.1
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
jsonschema==2.6.0
lru-dict==1.1.6
lxml==4.2.1
multidict==4.3.1
mysqlclient==1.3.12
netdisco==1.5.0
netifaces==0.10.7
oauthlib==2.1.0
parsimonious==0.8.0
parso==0.2.1
pbkdf2==1.3
pexpect==4.6.0
pickleshare==0.7.4
pkg-resources==0.0.0
pkginfo==1.4.2
plyvel==1.0.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py-ecc==1.4.2
py-evm==0.2.0a18
py-solc==2.2.0
pycparser==2.18
pycryptodome==3.6.3
pyethash==0.1.27
Pygments==2.2.0
PyMySQL==0.8.1
pysha3==1.0.2
python-dateutil==2.7.3
python-twitter==3.4.2
pytz==2018.4
PyYAML==3.12
repoze.lru==0.7
requests==2.19.1
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
rlp==0.6.0
scrypt==0.8.6
semantic-version==2.6.0
simplegeneric==0.8.1
six==1.11.0
toolz==0.9.0
tqdm==4.23.4
traitlets==4.3.2
trie==1.3.7
twine==1.11.0
uPnPClient==0.0.8
urllib3==1.23
wcwidth==0.1.7
web3==4.4.1
websockets==5.0.1
yarl==1.2.0
zeroconf==0.20.0
